requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlalchemy
aiosqlite
jinja2
python-dotenv
faiss-cpu
sentence-transformers
numpy
pandas
openpyxl
python-multipart
pyyaml
openai
rank-bm25
langchain-community
nltk
transformers

